shcnt: fix hanging mot_cnt thread
The mot_cnt thread hang if the requested target position was equal
to the current position. In this case the thread stopped the movement
and acted like a new requested interrupted ongoing movement: reevaluated
the target, stopped the movement again, and acted like a new request
interrupted ongoing movement...

Instead of acting like a new request interrupts the ongoing movement,
when the target equals the current position, the thread acts like the
movent timed out (is finished).